+++ utils/helpers.ts
@@ -1,8 +1,9 @@
 import { compile } from 'handlebars';
 import { Context } from '../common/notification/types';
 import { getLogger } from '../common/logger/logger';
+import { AchievementContextType } from '../common/achievement/types';
 
-export const renderTemplate = (template: string, context: Partial<Context>, strict = false) => {
+export const renderTemplate = (template: string, context: Partial<Context> | AchievementContextType, strict = false) => {
     const log = getLogger('Template Rendering');
     if (!template) {
         log.error('Template string undefined', new Error('Template string undefined'));
